updated tests. prepared for precision change on assignment.

+++ tests/regression/15-floats/04-branch-function.c
@@ -1,10 +1,10 @@
-// PARAM: --result pretty
+// PARAM: --result pretty --propset float_domain interval
 #include<assert.h>
 #include<stdlib.h>
 
 #define NUM 5.5
 
-int getNumber(){
+float getNumber(){
 	return NUM;
 }
